climate-advisor/service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
loguru==0.7.2
